Improvements for FPS independent effects
Current state of implementing dynamic FPS.
FPS can be changed by using the $fps chat command, e.g. "$fps 60".

Most effects scale correctly now. Still have to fix places where it's '*=' each frame... need time to think about a nice solution for them.

+++ Source Main 5.2/source/GM3rdChangeUp.cpp
@@ -474,7 +474,7 @@ bool SEASON3A::CGM3rdChangeUp::AttackEffectBalgasBarrackMonster(CHARACTER* c, OB
     switch (o->Type)
     {
     case MODEL_MONSTER01 + 91:
-        if (c->AttackTime == 14)
+        if ((int)c->AttackTime == 14)
         {
             CreateEffect(MODEL_ARROW_HOLY, o->Position, o->Angle, o->Light, 1, o, o->PKKey);
             return true;
