+++ Source Main 5.2/source/GMBattleCastle.cpp
@@ -1707,7 +1707,7 @@ namespace battleCastle
         switch (c->MonsterIndex)
         {
         case 104:
-            if (c->AttackTime == 5)
+            if ((int)c->AttackTime == 5)
             {
                 VectorCopy(o->Position, Position);
                 Position[2] += 500.f;
