+++ Source Main 5.2/source/ZzzOpenglUtil.h
@@ -9,7 +9,7 @@ extern unsigned int WindowHeight;
 extern vec3_t CollisionPosition;
 extern double  FPS;
 extern double  FPS_AVG;
-extern double  FPS_ANIMATION_FACTOR;
+extern float  FPS_ANIMATION_FACTOR;
 #if defined(LDS_FOR_DEVELOPMENT_TESTMODE) || defined(LDS_UNFIXED_FIXEDFRAME_FORDEBUG)		// 실FPS.(고정 20FPS 상황에서 추정용.)
 extern float g_fFrameEstimate;
 #endif // defined(LDS_FOR_DEVELOPMENT_TESTMODE) || defined(LDS_UNFIXED_FIXEDFRAME_FORDEBUG)
